feat: add endpoint to get all users
Add an API endpoint that will fetch all stored users in the database.
It's a GET request, a middleware was added to prevent the use of
other request methods from making request through this endpoint.

A 405 status code will be returned if a method other than GET is
used to get data, and 200 status code will be returned if the
request is successful.