feat: add endpoint to delete user
Add an API to delete a given user specified by his/her email. The
email is added as part of the url query.

A 200 status code is returned if delete operation is successful, while
a 400 status code is returned either if a user does not exist or the
email is invalid.