fix: await next action
when a request is made it returns the result and a warning that says
"API resolved without sending a response for <insert route>,
this may result in stalled requests."

This happens because the await keyword isn't used to invoke
the next() action in a middleware used to check the REST method
used on a given route.
By adding await, the warning no longer show up.