refactor: refactor authentication
Added firstname and lastname to user registration requirements, removed
name.

Created a middleware to check for and verify the presence of
token.

Used Jose for token generation and verification. Had to ditch
jsonwebtoken becuase it doesn't work in nextjs edge functions
used in middlewares. With this  change presence and verification
of token is done in the middleware and moved out of the REST logic.

A route which unauthenticated users making requests to authenticated
route will be directed to  was added.