refactor: set cookies on response headers
When users authenticate, either through login or new user registration,
cookies is being set on the response headers.

Also, in the middlware, the set cookie is being used as a ticket to get
access to protected resource.

Cookies as a means of security is chosen because it is accessible to both
the client and server. Furthermore, it is a little more secure on the
client when compared to the alternative; localStorage.