feat: implement password reset
Password reset was implemented. Users' who forgot their password
can use this feature to request for a change of password.

When a user clicks on the link to reset password, they're
taken to a page where they enter the email registered on the
app, a link is sent to the email which they can use to
reset the password.

Completes [HOM-4] [HOM-53]