chore: implement idle timer
Idle timer that logs out a user when idle for some certain
duration of time was added.

When a user stop interacting with the application he/she is deemed
idle. For this implementation, when a user is idle for 5 minutes
he/she is logged out, and will need to login again to access
authenticated routes.

Completes [HOM-60]