feat: show user listing stat
User's listing statistics were implemented to show the performance
of the listings. The following will be displayed on the user dashboard.
1. Total listings
2. Rented listings
3. Total income till date
4. Bar chart showing monthly earnings in a year
Bar chart showing month listings in a year.

All of these gives the user a glimpse of what they're doing
and gives them an idea of what they need to do to improve their
earnings.