feat: implement user logout
User logout was implemented making use endpoint from the server

When a user logs out, the access token is invalidated and
the refresh token is also invalidated and non-reusable.

This ensures that even if the access token is not expired as
at when a user logs out, it can't be used by an attacker, and
the refresh token cannot be used to generate a new access token.